requirements.txt
asgiref==3.7.2
Babel==2.14.0
beautifulsoup4==4.12.3
cffi==1.16.0
cryptography==42.0.5
distlib==0.3.8
Django==4.2.11
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-environ==0.11.2
django-extensions==3.2.3
django-jquery==3.1.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-pandas==0.6.6
django-redis==5.4.0
django-select2==8.2.0
django-tables2==2.7.0
django_adsense_injector==0.2.1
docxcompose==1.4.0
docxtpl==0.16.7
filelock==3.13.1
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.0
peewee==3.17.1
platformdirs==4.2.0
pycparser==2.21
pylokit==0.8.1
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
pytz==2024.1
redis==5.0.8
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
Werkzeug==3.0.1
